+++ tests/tests_pytorch/trainer/test_trainer.py
@@ -523,7 +523,6 @@ def test_trainer_max_steps_and_epochs_validation(max_epochs, max_steps, incorrec
         (None, 0, True, None),
         (0, -1, True, 0),
         (-1, 0, True, -1),
-        (0, -1, True, 0),
     ],
 )
 def test_trainer_max_steps_and_epochs_fit_loop_done(max_epochs, max_steps, is_done, correct_trainer_epochs):
@@ -1915,7 +1914,6 @@ def test_detect_anomaly_nan(tmpdir):
             2,
         ),
         pytest.param({"strategy": DDPStrategy()}, DDPStrategy, CPUAccelerator, 1, marks=RunIf(mps=False)),
-        ({"strategy": DDPStrategy(), "accelerator": "cuda", "devices": 2}, DDPStrategy, CUDAAccelerator, 2),
         (
             {"strategy": "ddp_spawn", "accelerator": "cuda", "devices": 2, "num_nodes": 2},
             DDPStrategy,
@@ -1981,7 +1979,6 @@ def test_dataloaders_are_not_loaded_if_disabled_through_limit_batches(running_st
     ("trainer_kwargs", "expected_device_ids"),
     [
         ({}, [0]),
-        ({"devices": 1}, [0]),
         ({"devices": 1}, [0]),
         ({"devices": "1"}, [0]),
         pytest.param({"devices": 2}, [0, 1], marks=RunIf(mps=False)),